feat(test): add payment flow e2e tests
